tiltfile: global yaml tests to their own file cuz they're getting big
